Use JSON file relative directory for the generated class package Refactor unit tests
