+++ apollo-compiler/src/test/graphql/com/apollostack/compiler/query/HeroDetailsExpected.java
@@ -1,0 +1,29 @@
+package com.apollostack.compiler.query;
+
+import java.lang.Integer;
+import java.lang.String;
+import java.util.List;
+
+public interface HeroDetails {
+  Character hero();
+
+  interface Character {
+    String name();
+
+    CharacterFriendsConnection friendsConnection();
+  }
+
+  interface CharacterFriendsConnection {
+    Integer totalCount();
+
+    List<CharacterFriendsConnectionFriendsEdge> edges();
+  }
+
+  interface CharacterFriendsConnectionFriendsEdge {
+    CharacterFriendsConnectionFriendsEdgeCharacter node();
+  }
+
+  interface CharacterFriendsConnectionFriendsEdgeCharacter {
+    String name();
+  }
+}
