+++ apollo-compiler/src/test/graphql/com/apollostack/compiler/query/HeroName.json
