+++ apollo-compiler/src/test/graphql/com/apollostack/compiler/query/HeroNameExpected.java
@@ -1,0 +1,11 @@
+package com.apollostack.compiler.query;
+
+import java.lang.String;
+
+public interface HeroName {
+  Character hero();
+
+  interface Character {
+    String name();
+  }
+}
