+++ apollo-compiler/src/test/graphql/com/apollostack/compiler/query/TwoHeroes.json
