+++ apollo-compiler/src/test/graphql/com/apollostack/compiler/query/TwoHeroesExpected.java
@@ -1,0 +1,13 @@
+package com.apollostack.compiler.query;
+
+import java.lang.String;
+
+public interface TwoHeroes {
+  Character r2();
+
+  Character luke();
+
+  interface Character {
+    String name();
+  }
+}
