+++ apollo-compiler/src/test/graphql/com/apollostack/compiler/query/TwoHeroesUniqueExpected.java
@@ -1,0 +1,19 @@
+package com.apollostack.compiler.query;
+
+import java.lang.String;
+
+public interface TwoHeroesUnique {
+  R2Character r2();
+
+  LukeCharacter luke();
+
+  interface LukeCharacter {
+    long id();
+
+    String name();
+  }
+
+  interface R2Character {
+    String name();
+  }
+}
